+++ src/test/kotlin/com/nextstep/jngcii/step2/RequestsTest.kt
@@ -1,19 +1,17 @@
 package com.nextstep.jngcii.step2
 
-import com.nextstep.jngcii.step2.vo.Number
-import com.nextstep.jngcii.step2.vo.Operator
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
 
-class InputTest {
+class RequestsTest {
     @Test
     fun `Input 생성 성공 테스트`() {
         val numbers = listOf(
-            Number("1"),
-            Number("2"),
-            Number("3"),
-            Number("4")
+            POperator("1"),
+            POperator("2"),
+            POperator("3"),
+            POperator("4")
         )
         val operators = listOf(
             Operator.MINUS,
@@ -21,15 +19,15 @@ class InputTest {
             Operator.DIVIDE
         )
 
-        assertDoesNotThrow { Input(numbers, operators) }
+        assertDoesNotThrow { Requests(numbers, operators) }
     }
 
     @Test
     fun `Input 생성 실패 테스트`() {
         val numbers = listOf(
-            Number("1"),
-            Number("2"),
-            Number("3")
+            POperator("1"),
+            POperator("2"),
+            POperator("3")
         )
         val operators = listOf(
             Operator.MINUS,
@@ -37,6 +35,6 @@ class InputTest {
             Operator.DIVIDE
         )
 
-        assertThrows<IllegalArgumentException>("잘못된 계산식입니다.") { Input(numbers, operators) }
+        assertThrows<IllegalArgumentException>("잘못된 계산식입니다.") { Requests(numbers, operators) }
     }
 }
